move createPacketBuffer and parsePacketData functions to serializer, also move protocol's exports to serializer
